# source/app/swagger.yml
openapi: 2.0
info:
  version: "1.1.0"
  title: "Orkestro | Ambiente Administrativo (Swagger)"
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /api/menus:
    get:
      tags:
        - "Cardápios"
      summary: "Todos os Cardápios"
      description: "Fetch a list of all available menus."
      responses:
        '200':
          description: "A list of menus"
    post:
      tags:
        - "Cardápios"
      summary: "Criar Cardápio"
      description: "Criar um novo cardápio"
      parameters:
        - in: body
          name: menu
          description: "Menu object to be created"
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
              description:
                type: string
      responses:
        '201':
          description: "Menu created successfully"
  /api/menus/{menu_id}:
    delete:
      tags:
        - "Cardápios"
      summary: "Deletar Cardápio"
      description: "Deletar cardápio através do identificador (UUID)"
      parameters:
        - in: body
    put:
      tags:
        - "Cardápios"
      summary: "Atualizar Cardápio"
      description: "Atualizar cardápios através do identificador (UUID)"
      parameters:
        - in: body
...
  /api/menus/{menu_id}/categories:
    get:
      tags:
        - "Categorias"
      summary: "Todas as Categorias"
      description: "Fetch a list of all available categories."
      responses:
        '200':
          description: "A list of categories"
    post:
      tags:
        - "Categorias"
      summary: "Criar Categoria"
      description: "Add a new category to the system."
      parameters:
        - in: body
          name: category
          description: "Category object to be created"
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
              description:
                type: string
      responses:
        '201':
          description: "Category created successfully"
...